:recycle: Move secret messages service utilities and filters to separate files
