+++ src/services/__init__.py
@@ -1,5 +1,7 @@
 from .contacts import *
+from .filters import *
 from .food_menu import *
 from .notifiers import *
 from .photos import *
+from .secret_messages import *
 from .work import *
