+++ src/services/secret_messages.py
@@ -1,0 +1,28 @@
+from typing import Iterable
+
+from models import Contact, HasUserId
+
+__all__ = (
+    'can_see_team_secret',
+    'can_see_contact_secret',
+)
+
+
+def can_see_team_secret(
+        *,
+        user_id: int,
+        team_members: Iterable[HasUserId],
+) -> bool:
+    user_ids = {member.user_id for member in team_members}
+    return user_id in user_ids
+
+
+def can_see_contact_secret(
+        *,
+        user_id: int,
+        contact: Contact,
+) -> bool:
+    return user_id in (
+        contact.of_user.id,
+        contact.to_user.id,
+    )
